+++ library/psr11/Container.php
@@ -7,6 +7,31 @@ use suda\Registry as SudaContainer;
 
 class Container extends SudaContainer implements ContainerInterface
 {
+    function withPsrContainer(ContainerInterface $container) {
+        return $this->withDelegate(new class($container) extends SudaContainer {
+            /** @var ContainerInterface */
+            private $container;
+
+            function __construct(ContainerInterface $container) {
+                $this->container = $container;
+            }
+
+            function offsetGet($key) {
+                if ($this->container->has($key)) {
+                    return $this->container->get($key);
+                }
+                return parent::offsetGet($key);
+            }
+
+            function offsetExists($key) {
+                if ($this->container->has($key)) {
+                    return true;
+                }
+                return parent::offsetExists($key);
+            }
+        });
+    }
+
     function get($id) {
         try {
             $ret = $this[$id];